# Conflicts:
#	bibitem_parsing/bibitem_parser.py
#	texparser/information_extractor.py

+++ parser_bot.py
@@ -9,7 +9,8 @@ from texparser.tar_extractor import TarExtractor
 
 
 class ParserBot:
-    def __init__(self, dataset_folder_path, extract_folder_path: str = "") -> None:
+    def __init__(self, file_to_start: str, dataset_folder_path, extract_folder_path: str = "") -> None:
+        self.file_to_start = file_to_start
         self.dataset_folder_path = dataset_folder_path
         self.extract_folder_path = extract_folder_path
         self.tarExtractor = TarExtractor(
@@ -19,22 +20,26 @@ class ParserBot:
     def run(self):
         self.tarExtractor.create_extract_folder_path()
         filenames = os.listdir(self.dataset_folder_path)
+        found_start_file = False
         for count, filename in enumerate(tqdm(filenames)):
-            try:
-                self.tarExtractor.untar_file_into_folder(
-                    self.dataset_folder_path + filename)
-            except Exception as e:
-                print(e)
-            extract_folder = self.extract_folder_path + \
-                             self.dataset_folder_path + filename.replace(".tar", "")
-            sub_extract_folder = extract_folder + \
-                                 "/" + os.listdir(extract_folder)[0]
-            self.tarExtractor.extract_folder(
-                sub_extract_folder, self.tarExtractor.untargz_file_into_folder)[1]
-            data, author_citation_tuples, author_citation_tuples_failed = self.informationExtractor.extract_all(
-                sub_extract_folder)
-            self.log_progress(data, author_citation_tuples, author_citation_tuples_failed, count)
-            shutil.rmtree(extract_folder)
+            if filename == self.file_to_start:
+                found_start_file = True
+            if found_start_file:
+                try:
+                    self.tarExtractor.untar_file_into_folder(
+                        self.dataset_folder_path + filename)
+                except Exception as e:
+                    print(e)
+                extract_folder = self.extract_folder_path + \
+                                self.dataset_folder_path + filename.replace(".tar", "")
+                sub_extract_folder = extract_folder + \
+                                    "/" + os.listdir(extract_folder)[0]
+                self.tarExtractor.extract_folder(
+                    sub_extract_folder, self.tarExtractor.untargz_file_into_folder)[1]
+                data, author_citation_tuples, author_citation_tuples_failed = self.informationExtractor.extract_all(
+                    sub_extract_folder)
+                self.log_progress(data, author_citation_tuples, author_citation_tuples_failed, count)
+                shutil.rmtree(extract_folder)
         self.tarExtractor.delete_extract_folder_path()
 
     def log_progress(self, data: dict, author_citation_tuples, author_citation_tuples_failed, count: int) -> None:
@@ -50,6 +55,6 @@ class ParserBot:
 
 
 if __name__ == "__main__":
-    parser = ParserBot("content/", "extract/")
+    parser = ParserBot("arXiv_src_1705_015.tar", "content/", "extract/")
     print("Parser started")
     parser.run()
